+++ components/src/main/java/org/patternfly/component/card/CardHeader.java
@@ -34,7 +34,7 @@ import static org.jboss.elemento.Elements.span;
 import static org.jboss.elemento.EventType.click;
 import static org.patternfly.component.button.Button.button;
 import static org.patternfly.core.Aria.expanded;
-import static org.patternfly.icon.IconSets.rhUi.caretRight;
+import static org.patternfly.icon.IconSets.rhMicrons.caretDown;
 import static org.patternfly.style.Classes.component;
 import static org.patternfly.style.Classes.header;
 import static org.patternfly.style.Classes.icon;
@@ -102,7 +102,7 @@ public class CardHeader extends CardSubComponent<HTMLDivElement, CardHeader> imp
                             })
                             .on(click, e -> card.toggle())
                             .add(span().css(component(Classes.card, header, toggle, icon))
-                                    .add(caretRight().element())))
+                                    .add(caretDown().element())))
                     .element();
             if (card.toggleRightAligned) {
                 css(modifier(toggleRight));